+++ src/App.tsx
@@ -4,24 +4,31 @@ import { LiveRanking } from "./components/LiveRanking";
 import { NoData } from "./components/NoData";
 import { PageFooter } from "./components/PageFooter";
 import { useLocation } from "react-router-dom";
+import { ActiveCompList } from "./components/CompList";
+import { CompList } from "./types/CompList";
 
 function App() {
-  const search = useLocation().search;
-  const id = new URLSearchParams(search).get("id");
+  const [liveData, setLiveData] = useState();
+  const [taskData, setTaskData] = useState();
+  const [compList, setCompList] = useState<CompList | undefined>();
+  const [failedFetchAttempts, setFailedFetchAttempts] = useState(0);
 
-  // const BASE_URL = "https://race.airtribune.com/";
-  const BASE_URL = "https://corsproxy.io/?https://race.airtribune.com/";
+  const search = useLocation().search;
+  const compId = new URLSearchParams(search).get("id");
+  const useCorsProxy = import.meta.env.VITE_USE_CORS_PROXY === "true";
+
+  const BASE_URL = useCorsProxy
+    ? "https://corsproxy.io/?https://race.airtribune.com/"
+    : "https://race.airtribune.com/";
   // const BASE_URL = "http://127.0.0.1:5500/src/demo-data/";
 
   const REFRESH_INTERVAL_TASK = 36_000_000; // in ms
   const REFRESH_INTERVAL_LIVE = 3000; // in ms
 
-  const [liveData, setLiveData] = useState();
-  const [taskData, setTaskData] = useState();
-  const [failedFetchAttempts, setFailedFetchAttempts] = useState(0);
+  const LIVE_TASK_URL = BASE_URL + compId + "/feed_task.json";
+  const LIVE_DATA_URL = BASE_URL + compId + "/feed_live.json";
+  const COMP_LIST_URL = BASE_URL + "/active_comps.json";
 
-  const LIVE_TASK_URL = BASE_URL + id + "/feed_task.json";
-  const LIVE_DATA_URL = BASE_URL + id + "/feed_live.json";
   const showConnectionWarning = failedFetchAttempts > 2;
 
   const fetchLiveData = async () => {
@@ -59,7 +66,7 @@ function App() {
   };
 
   useEffect(() => {
-    if (!taskData) fetchTaskData();
+    if (!taskData && compId) fetchTaskData();
     fetchLiveData();
     const fetchInterval = setInterval(fetchLiveData, REFRESH_INTERVAL_LIVE);
     const fetchIntervalTask = setInterval(fetchTaskData, REFRESH_INTERVAL_TASK);
@@ -69,18 +76,34 @@ function App() {
     };
   }, [taskData]);
 
-  if (!id) return <h4>Let me check what to show here</h4>;
-  if (!taskData) return <NoData />;
+  const fetchCompList = async () => {
+    try {
+      const res = await fetch(COMP_LIST_URL);
+      setCompList(await res.json());
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  useEffect(() => {
+    fetchCompList();
+  }, []);
 
   return (
-    <div className="flex flex-col min-h-screen dark:bg-slate-800 ">
-      <div className="flex-grow">
-        <TaskInfo
-          taskData={taskData}
-          showConnectionWarning={showConnectionWarning}
-        />
-        <hr className="dark:border-slate-500" />
-        {!liveData ? <NoData /> : <LiveRanking liveData={liveData} />}
+    <div className="flex flex-col min-h-screen dark:bg-slate-800  ">
+      <div className="flex-grow  text-gray-700 dark:text-slate-400 text-sm even:bg-neutral-100 dark:even:bg-slate-700">
+        {!compId && <ActiveCompList compList={compList} />}
+        {compId && !taskData && <NoData />}
+        {compId && taskData && (
+          <>
+            <TaskInfo
+              taskData={taskData}
+              showConnectionWarning={showConnectionWarning}
+            />
+            <hr className="dark:border-slate-500" />
+            {!liveData ? <NoData /> : <LiveRanking liveData={liveData} />}
+          </>
+        )}
       </div>
       <PageFooter />
     </div>