+++ src/components/NoData.tsx
@@ -1,7 +1,5 @@
 export function NoData() {
   return (
-    <div className="text-center m-10 text-lg text-gray-800">
-      No data available yet 😕
-    </div>
+    <div className="text-center m-10 text-lg ">No data available yet 😕</div>
   );
 }