+++ src/types/CompList.ts
@@ -1,0 +1,3 @@
+export interface CompList {
+  activeComps?: { name?: string; id?: number }[];
+}